+++ migrations/CHANGELOG.md
@@ -3,3 +3,9 @@
 ## Version 1
 
 - Initial Database Schema
+
+## Version 2 - Implement Dynamic Questionnaires
+
+- Add tables preguntas, cuestionarios, preguntas_cuestionarios, antecedentes, consultas_preguntas
+
+- Drop column Pacientes -> Antecedentes, Consultas -> Evaluacion